@@ -1,59 +1,53 @@
 ﻿using Application.DTO;
 using Application.SeedWorks;
 using Core.Contacts;
-using MediatR;
-using Repositories.Unit;
-using System;
-using System.Threading;
-using System.Threading.Tasks;
 
-namespace Application.Contacts
+namespace Application.Contacts;
+
+public class Create
 {
-    public class Create
+    public class Command : IRequest<Result<Unit>>
     {
-        public class Command : IRequest<Result<Unit>>
-        {
-            public ContactFormDTO ContactForm { get; set; }
+        public ContactFormDTO ContactForm { get; set; }
 
+    }
+
+    public class CommandHandler : IRequestHandler<Command, Result<Unit>>
+    {   
+        private readonly UnitWrapper _context;
+        
+        public CommandHandler(UnitWrapper context)
+        {
+            _context = context;
         }
 
-        public class CommandHandler : IRequestHandler<Command, Result<Unit>>
-        {   
-            private readonly UnitWrapper _context;
+        public async Task<Result<Unit>> Handle(Command request, 
+            CancellationToken cancellationToken)
+        {
+
+            var contact = Contact.Create(
+                request.ContactForm.Title,
+                request.ContactForm.FirstName,
+                request.ContactForm.MiddleName,
+                request.ContactForm.LastName,
+                request.ContactForm.Gender,
+                request.ContactForm.MobileNo,
+                request.ContactForm.EmailAddress,
+                request.ContactForm.PrimaryContact,
+                request.ContactForm.Location,
+                request.ContactForm.GroupTag);
             
-            public CommandHandler(UnitWrapper context)
+            try
             {
-                _context = context;
+                 _context.Customers.Add(contact);
+                await _context.SaveChangesAsync();
             }
-
-            public async Task<Result<Unit>> Handle(Command request, 
-                CancellationToken cancellationToken)
+            catch (Exception ex)
             {
-
-                var contact = Contact.Create(
-                    request.ContactForm.Title,
-                    request.ContactForm.FirstName,
-                    request.ContactForm.MiddleName,
-                    request.ContactForm.LastName,
-                    request.ContactForm.Gender,
-                    request.ContactForm.MobileNo,
-                    request.ContactForm.EmailAddress,
-                    request.ContactForm.PrimaryContact,
-                    request.ContactForm.Location,
-                    request.ContactForm.GroupTag);
-                
-                try
-                {
-                     _context.Customers.Add(contact);
-                    await _context.SaveChangesAsync();
-                }
-                catch (Exception ex)
-                {
-                    return Result<Unit>.Failure(ex.Message);
-                }
-                
-                return Result<Unit>.Success(Unit.Value);
+                return Result<Unit>.Failure(ex.Message);
             }
+            
+            return Result<Unit>.Success(Unit.Value);
         }
     }
 }
