@@ -1,35 +1,69 @@
-﻿using Application.Campaigns.Queries;
-using Application.Configuration;
-using FluentValidation;
-using MediatR;
+﻿using Application.Configuration;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Persistence.Context;
-using Repositories.Unit;
 using System.Reflection;
 
-namespace Application.Extensions
-{
-    public static class ApplicationServiceExtension
-    {
-        public static IServiceCollection AddApplicationServices(this IServiceCollection services,
-            IConfiguration configuration)
-        {
+namespace Application.Extensions;
 
-            services.AddMediatR(typeof(List.QueryHandler));
+public static class ApplicationServiceExtension
+{
+    public static IServiceCollection AddApplicationServices(this IServiceCollection services,
+        IConfiguration configuration)
+    {
+
+
+        services.AddMediatR(opt =>
+        {
+            opt.RegisterServicesFromAssembly(Assembly.GetExecutingAssembly());
+        });
+
+        services.AddDbContext<DataContext>(opt => {
+             var env = Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT");
+
+            string connStr;
+
             
-            services.AddDbContext<DataContext>(opt =>
+            // Depending on if in development or production, use either FlyIO
+            // connection string, or development connection string from env var.
+            if (env == "Development")
             {
-                //opt.UseSqlite(configuration.GetConnectionString("SqliteConnection"));
-                opt.UseNpgsql(configuration.GetConnectionString("PostgreConnection"));
-            });
-            services.AddValidatorsFromAssembly(Assembly.GetExecutingAssembly());
-            services.AddAutoMapper(typeof(MappingProfiles));
-            //repository
-            services.AddScoped<UnitWrapper>();
-            
-            return services;
-        }
+                // Use connection string from file.
+                connStr = configuration.GetConnectionString("PostgreConnection");
+            }
+            else
+            {
+                // Use connection string provided at runtime by FlyIO.
+                var connUrl = Environment.GetEnvironmentVariable("DATABASE_URL");
+
+                // Parse connection URL to connection string for Npgsql
+                connUrl = connUrl.Replace("postgres://", string.Empty);
+                var pgUserPass = connUrl.Split("@")[0];
+                var pgHostPortDb = connUrl.Split("@")[1];
+                var pgHostPort = pgHostPortDb.Split("/")[0];
+                var pgDb = pgHostPortDb.Split("/")[1];
+                var pgUser = pgUserPass.Split(":")[0];
+                var pgPass = pgUserPass.Split(":")[1];
+                var pgHost = pgHostPort.Split(":")[0];
+                var pgPort = pgHostPort.Split(":")[1];
+                var updatedHost = pgHost.Replace("flycast", "internal");
+
+                connStr = $"Server={updatedHost};Port={pgPort};User Id={pgUser};Password={pgPass};Database={pgDb};";
+            }
+
+            // Whether the connection string came from the local development configuration file
+            // or from the environment variable from FlyIO, use it to set up your DbContext.
+            opt.UseNpgsql(connStr);
+        });
+
+        
+
+        services.AddValidatorsFromAssembly(Assembly.GetExecutingAssembly());
+        services.AddAutoMapper(typeof(MappingProfiles));
+        //repository
+        services.AddScoped<UnitWrapper>();
+        
+        return services;
     }
 }
