@@ -6,7 +6,7 @@ using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Threading.Tasks;
 
-namespace API.Controllers
+namespace WCMAPI.Controllers
 {
 
     [Route("api/[controller]")]
