@@ -9,7 +9,7 @@ using System.Security.Claims;
 using System.Security.Cryptography;
 using System.Text;
 
-namespace API.Services
+namespace WCMAPI.Services
 {
     public class TokenService
     {
