@@ -1,5 +1,5 @@
-using API.Extensions;
-using API.Middleware;
+using WCMAPI.Extensions;
+using WCMAPI.Middleware;
 using Application.Extensions;
 using Infrastructure;
 using Microsoft.AspNetCore.Builder;
@@ -8,7 +8,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
-namespace API
+namespace WCMAPI
 {
     public class Startup
     {
