@@ -1,20 +1,11 @@
 ﻿using Application.SeedWorks;
 using Core.Enum;
-using MediatR;
-using Repositories.Unit;
-using System;
-using System.Threading;
-using System.Threading.Tasks;
 
 namespace Application.Activities.Commands
 {
     public class DeleteActivity
     {
-
-        public class Command : IRequest<Result<Unit>>
-        {
-            public string ActivityId { get; set; }
-        }
+        public record Command(string ActivityId) : IRequest<Result<Unit>>;
 
         public class CommandHandler : IRequestHandler<Command, Result<Unit>>
         {
@@ -31,13 +22,6 @@ namespace Application.Activities.Commands
                     var activity = await _context.CampaignRepo.GetByActivity(request.ActivityId);
                     if (activity == null)
                         throw new Exception("Activity not found");
-
-
-                    //check if the activity is pending
-                    if(activity.Status != ActivityStatusEnum.Pending.ToString())
-                    {
-                        throw new Exception("Activity is no longer pending and cannot be deleted");
-                    }
 
                     _context.CampaignRepo.DeleteActivity(activity);
                     
