@@ -9,7 +9,7 @@ using System.Linq;
 namespace Application.DTOs
 {
 
-    public record  MarketingSubTaskDTO(string Id,string Task,string AssignedTo,string AssignedBy, string Status,bool MarkDelete = false);
+    public record  MarketingSubTaskDTO(string Id,string Task,string AssignedTo,string? AssignedBy, string? Status,bool MarkDelete = false);
 
     public class MarketingTaskDTO {
         public Guid Id { get; set; }
@@ -17,9 +17,18 @@ namespace Application.DTOs
         public string UserName { get; set; } = "";
         public bool Close { get; set; } = false;
 
-        
-        public int TotalSubTaskCount => SubTasks.Where(s => s.Status != TaskStatusEnum.Done.ToString()).Count();
-        public List<MarketingSubTaskDTO> SubTasks { get; set; }
+
+        public int TotalSubTaskCount
+        {
+            get
+            {
+                var subTask = SubTasks.Where(s => s.Status != TaskStatusEnum.Done.ToString()).ToList();
+                return (subTask != null) ? subTask.Count() : 0;
+                
+            }
+        }
+
+        public List<MarketingSubTaskDTO> SubTasks { get; set; } = new();
         public List<string> Users => SubTasks.Select(s => s.AssignedTo).Distinct().ToList();
     }
 
