@@ -4,7 +4,7 @@
     {
         public bool IsSuccess { get; set; }
         public T Value { get; set; }
-        public string Error { get; set; }
+        public string Error { get; set; } = string.Empty;
 
         public static Result<T> Success(T value) => new Result<T> { IsSuccess = true, Value = value };
         public static Result<T> Failure(string error) => new Result<T> { IsSuccess = false, Error = error };
