@@ -1,11 +1,14 @@
-﻿using Application.DTO;
+﻿using Application.Configuration;
+using Application.DTO;
 using Application.Interface;
 using Core.Campaigns;
 using Core.Channels;
 using Infrastructure.External.Email;
 using Infrastructure.External.SMS;
 using Infrastructure.External.Web;
+using Microsoft.Extensions.Options;
 using Repositories.Unit;
+using System.Net.Http;
 
 namespace Infrastructure.Services.Campaigns
 {
@@ -16,21 +19,27 @@ namespace Infrastructure.Services.Campaigns
         private readonly IWebPostService webPostService;
         private readonly ISMSService smsService;
         private readonly ISendSMTPClient sendSMTPClient;
+        private readonly IOptions<TwilioOption> options;
+        private readonly HttpClient _httpClient;
 
         public ServiceFactory(
             IWebPostService webPostService,
             ISMSService smsService,
-            ISendSMTPClient sendSMTPClient)
+            ISendSMTPClient sendSMTPClient,
+            IOptions<TwilioOption> options,
+            HttpClient httpClient)
         {
             
            
             this.webPostService = webPostService;
             this.smsService = smsService;
             this.sendSMTPClient = sendSMTPClient;
+            this.options = options;
+            _httpClient = httpClient;
         }
 
         public IActivityServiceAccessor GetSMS(ChannelSetting setting) {
-            return new SMSActivityService(smsService, setting);
+            return new SMSActivityService(smsService, setting,options);
         }
 
         public IActivityServiceAccessor GetWebPost(ChannelSetting setting)
@@ -45,7 +54,7 @@ namespace Infrastructure.Services.Campaigns
 
         public IActivityServiceAccessor GetSocial(ChannelSetting setting)
         {
-            return new FacebookService(setting);
+            return new SocialActivityService(_httpClient,setting);
         }
     }
 }
